refactor: Changes the names of few css files to maintain naming consistency
