+++ src/Pages/index.js
@@ -1,4 +1,4 @@
-export { Landing } from "./Landing/Landing-page";
+export { Landing } from "./Landing/LandingPage";
 export { SignUp } from "./Authentication/Signup";
 export { Login } from "./Authentication/Login";
 export { Home } from "./Home/Home";
